[_compilers] factor out shared code between interpolatable compilers
