+++ Lib/ufo2ft/_compilers/otfCompiler.py
@@ -4,7 +4,6 @@ from typing import Optional, Type
 from ufo2ft.constants import CFFOptimization
 from ufo2ft.outlineCompiler import OutlineOTFCompiler
 from ufo2ft.preProcessor import OTFPreProcessor
-from ufo2ft.util import prune_unknown_kwargs
 
 from .baseCompiler import BaseCompiler
 
@@ -17,15 +16,3 @@ class OTFCompiler(BaseCompiler):
     roundTolerance: Optional[float] = None
     cffVersion: int = 1
     subroutinizer: Optional[str] = None
-    _tables: Optional[list] = None
-    extraSubstitutions: Optional[dict] = None
-
-    def postprocess(self, font, ufo, glyphSet, info=None):
-        if self.postProcessorClass is not None:
-            postProcessor = self.postProcessorClass(
-                font, ufo, glyphSet=glyphSet, info=info
-            )
-            kwargs = prune_unknown_kwargs(self.__dict__, postProcessor.process)
-            font = postProcessor.process(**kwargs)
-        self._glyphSet = glyphSet
-        return font
